requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.3.2
boto3==1.38.8
botocore==1.38.8
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
Django==5.2
django-storages==1.14.6
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.0
hexbytes==1.3.0
idna==3.10
jmespath==1.0.1
multidict==6.4.3
parsimonious==0.10.0
propcache==0.3.1
psycopg2==2.9.10
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
web3==7.11.0
websockets==15.0.1
yarl==1.20.0
